Seprated the test, moved test.html.twig into tmp folder, test.config.adjusted for templates
