--- templates/test.html.twig
@@ -1,1 +1,0 @@
-Hello, {{ name }}!
